세션관리 기능 구현 완료

+++ omilzomil/backend/src/app/api/endpoints/military_unit.py
@@ -4,13 +4,19 @@ from sqlalchemy.orm import Session
 from app.crud import military_unit as crud
 from app.schemas import military_unit as schema
 from app.api import deps
+from app.schemas.user import UserReadResponse
 
 
 router = APIRouter()
 
 
 @router.post("/", response_model=schema.MilitaryUnitResponse)
-async def create_military_unit(unit: schema.MilitaryUnitCreate = Body(), db: Session = Depends(deps.get_db)):
+async def create_military_unit(
+    unit: schema.MilitaryUnitCreate = Body(), db: Session = Depends(deps.get_db), current_user: UserReadResponse = Depends(deps.get_current_active_super)
+):
+    if not current_user.success:
+        return schema.MilitaryUnitResponse(success=False, message=current_user.message)
+
     return crud.create_military_unit(db, unit.unit)
 
 
@@ -25,10 +31,21 @@ def get_military_unit(unit_id: int, db: Session = Depends(deps.get_db)):
 
 
 @router.put("/{unit_id}", response_model=schema.MilitaryUnitResponse)
-async def update_military_unit(unit_id: int, new_unit: schema.MilitaryUnitUpdate = Body(), db: Session = Depends(deps.get_db)):
+async def update_military_unit(
+    unit_id: int,
+    new_unit: schema.MilitaryUnitUpdate = Body(),
+    db: Session = Depends(deps.get_db),
+    current_user: UserReadResponse = Depends(deps.get_current_active_super),
+):
+    if not current_user.success:
+        return schema.MilitaryUnitResponse(success=False, message=current_user.message)
+
     return crud.update_military_unit(db, unit_id, new_unit.unit)
 
 
 @router.delete("/{unit_id}", response_model=schema.MilitaryUnitResponse)
-def delete_military_unit(unit_id: int, db: Session = Depends(deps.get_db)):
+def delete_military_unit(unit_id: int, db: Session = Depends(deps.get_db), current_user: UserReadResponse = Depends(deps.get_current_active_super)):
+    if not current_user.success:
+        return schema.MilitaryUnitResponse(success=False, message=current_user.message)
+
     return crud.delete_military_unit(db, unit_id)
